use consistent conditional checks...
...with a PATH check at the end of each installation script.

We shouldn't use `command -v mybinary` as our conditional check for
installing our binary as whether or not its on the PATH is not important
for the binary's existence. So if we have a PATH issue, we wouldn't have
to install our binary again for no reason. Now once it's installed, we
do check if it's on the PATH so we detect PATH issues as early as
possible.