bastion: add tf code for wacky idea
basically, to be able to access all of my homes, I can spin up a bastion
host, start a remote-forwarded SSH connections to it from A, and then
connect to A from B by proxying through the bastion using that forwarded
SSH connection. it's very silly and hacky, but very fun.